Исправил проблему отображения дочерних пакетов. Выделил функцию для получения узлов и ребер в отдельный файл utils/graphBuilder.js
